+++ src/UI/ledger/LedgerLoginContainer/types.ts
@@ -2,13 +2,13 @@ import type { AddressTablePropsType } from './AddressTable';
 import type { ConfirmAddressPropsType } from './ConfirmAddress';
 import type { LedgerConnectPropsType } from './LedgerConnect';
 import type { LedgerLoadingPropsType } from './LedgerLoading';
-import type { LedgerProgressBarPropsType } from './LedgerProgressBar';
+import type { LedgerProgressStepsPropsType } from './LedgerProgressSteps';
 
 export interface InnerLedgerComponentsClassesType {
   confirmAddressClassNames?: ConfirmAddressPropsType['confirmAddressClassNames'];
   addressTableClassNames?: AddressTablePropsType['addressTableClassNames'];
   ledgerLoadingClassNames?: LedgerLoadingPropsType['ledgerLoadingClassNames'];
   ledgerConnectClassNames?: LedgerConnectPropsType['ledgerConnectClassNames'];
-  ledgerProgressBarClassNames?: LedgerProgressBarPropsType['ledgerProgressBarClassNames'];
+  ledgerProgressStepsClassNames?: LedgerProgressStepsPropsType['ledgerProgressStepsClassNames'];
   ledgerScamPhishingAlertClassName?: string;
 }